fix issues with products.js
