+++ public/products.js
@@ -1,0 +1,1422 @@
+// src="grivna_20.jpg"
+const products = [
+    // No1
+    {
+        price      : 198,
+        name       : 'Сребърна мъжка гривна с кожа',
+        id         : '20GRV',
+        slug       : 'sreburna-mujka-grivna-s-koja',
+        oldSlug    : '/large/20grv.html',
+        desc       : '',
+        sku        : '20GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '20grv.jpg' // !!!!!!!!
+    },
+    // No2
+    {
+        price      : 179,
+        name       : 'Сребърна гривна, изработена от въже за катерене (алпинистко) с инкрустиран лазурит',
+        id         : '25GRV',
+        slug       : 'sreburna-grivna-izrabotena-ot-vuzhe-za-katerene',
+        oldSlug    : '/large/25grv.html',
+        desc       : '',
+        sku        : '25GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '25grv.jpg'
+    },
+    // No3
+    {
+        price      : 449,
+        name       : 'Ръчно изработена (кована) сребърна гривна',
+        id         : '21GRV',
+        slug       : 'ruchno-izrabotena-kovana-sreburna-grivna-inkrustiran-brilyant-sus-zlaten-obkov',
+        oldSlug    : '/large/21grv.html',
+        desc       : 'Инкрустиран брилянт със златен обков',
+        sku        : '21GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '21grv.jpg'
+    },
+    // No4
+    {
+        price      : 598,
+        name       : 'Ръчно изработена (кована) сребърна гривна;',
+        id         : '22GRV',
+        slug       : 'ruchno-izrabotena-kovana-sreburna-grivna',
+        oldSlug    : '/large/22grv.html',
+        desc       : 'инкрустиран брилянт със златен обков (14к); златни апликации',
+        sku        : '22GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '22grv.jpg'
+    },
+    // No5
+    {
+        price      : 279,
+        name       : 'Мъжка гривна от сребро и естествена кожа с ахат и черни циркони (2 бр.)',
+        id         : '23GRV',
+        slug       : 'muzhka-grivna-ot-srebro-i-estestvena-kozha-s-ahat-i-cherni-cirkoni-2-br',
+        oldSlug    : '/large/23grv.html ',
+        desc       : '',
+        sku        : '23GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '23grv.jpg'
+    },
+    // No6
+    {
+        price      : 449,
+        name       : 'Ръчно изработена (кована) сребърна гривна със златни апликации и инкрустиран брилянт със златен обков;',
+        id         : '24GRV',
+        slug       : 'ruchno-izrabotena-kovana-sreburna-grivna-sus-zlatni-aplikacii-i-inkrustiran-brilyant-sus-zlaten-obkov',
+        oldSlug    : '/large/24grv.html',
+        desc       : 'Ръчно изработена (кована) сребърна гривна със златни апликации и инкрустиран брилянт със златен обков; Злато 14к',
+        sku        : '24GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '24grv.jpg'
+    },
+    // No7
+    {
+        price      : 98,
+        name       : 'Сребро и естествена кожа (черна, кафява и бордо)',
+        id         : '1GRV',
+        slug       : 'srebro-i-estestvena-kozha-cherna-kafyava-i-bordo',
+        oldSlug    : '/large/large1.htm',
+        desc       : '',
+        sku        : '1GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '1grv.jpg'
+    },
+    // No8
+    {
+        price      : 189,
+        name       : 'Кована сребърна гривна с емайли от малахит и лазурит',
+        id         : '2GRV',
+        slug       : 'kovana-sreburna-grivna-s-emajli-ot-malahit-i-lazurit',
+        oldSlug    : '/large/large2.htm',
+        desc       : '',
+        sku        : '2GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '2grv.jpg'
+    },
+    // No9
+    {
+        price      : 219,
+        name       : 'Сребърна гривна с естествена кожа (черна или кафява)',
+        id         : '13GRV',
+        slug       : 'sreburna-grivna-s-estestvena-kozha-cherna-ili-kafyava',
+        oldSlug    : '/large/large13l.htm',
+        desc       : '',
+        sku        : '13GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '13grv.jpg'
+    },
+    // No10
+    {
+        price      : 469,
+        name       : 'Ръчно изработена (кована) сребърна гривна, инкрустирана със злато 14к',
+        id         : '16GRV',
+        slug       : 'ruchno-izrabotena-kovana-sreburna-grivna-inkrustirana-sus-zlato-14k',
+        oldSlug    : '/large/16gr_large.html',
+        desc       : '',
+        sku        : '16GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '16grv.jpg'
+    },
+    // No11
+    {
+        price      : 529,
+        name       : 'Ръчно изработена (кована) сребърна гривна, инкрустирана със злато 14к и брилянт - 0.03k-VS-FG',
+        id         : '17GRV',
+        slug       : 'ruchno-izrabotena-kovana-sreburna-grivna-inkrustirana-sus-zlato-14k-i-brilyant-003k-vs-fg',
+        oldSlug    : '/large/17gr_large.html',
+        desc       : '',
+        sku        : '17GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '17grv.jpg'
+    },
+    // No12
+    {
+        price      : 279,
+        name       : 'Ръчно изработена (кована) сребърна гривна;',
+        id         : '18GRV',
+        slug       : 'ruchno-izrabotena-kovana-sreburna-grivna',
+        oldSlug    : '/large/18gr_large.html',
+        desc       : 'Инкрустиран смарагд със златен обков',
+        sku        : '18GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '18grv.jpg'
+    },
+    // No13
+    {
+        price      : 268,
+        name       : 'Масивна сребърна гривна',
+        id         : '7GR',
+        slug       : 'masivna-sreburna-grivna',
+        oldSlug    : '/large/large7.htm',
+        desc       : '',
+        sku        : '7GR',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '7grv.jpg'
+    },
+    // No14
+    {
+        price      : 329,
+        name       : 'Сребро и естествена кожа (черна)',
+        id         : '8GRV',
+        slug       : 'srebro-i-estestvena-kozha-cherna',
+        oldSlug    : '/large/large8.htm',
+        desc       : '',
+        sku        : '8GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '8grv.jpg'
+    },
+    // No15
+    {
+        price      : 369,
+        name       : 'Сребърна гривна с хематит',
+        id         : '9GRV',
+        slug       : 'sreburna-grivna-s-hematit',
+        oldSlug    : '/large/large9.htm',
+        desc       : '',
+        sku        : '9GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '9grv.jpg'
+    },
+    // No16
+    {
+        price      : 129,
+        name       : 'Сребро и естествена кожа (черна, кафява и бордо)',
+        id         : '10GRV',
+        slug       : 'srebro-i-estestvena-kozha-cherna-kafyava-i-bordo',
+        oldSlug    : '/large/large10.htm',
+        desc       : '',
+        sku        : '10GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '10grv.jpg'
+    },
+    // No17
+    {
+        price      : 269,
+        name       : 'Гривна, изработена от огъната сребърна вилица и емайли',
+        id         : '11GRV',
+        slug       : 'grivna-izrabotena-ot-ogunata-sreburna-vilica-i-emajli',
+        oldSlug    : '/large/large11.htm',
+        desc       : '',
+        sku        : '11GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '11grv.jpg'
+    },
+    // No18
+    {
+        price      : 98,
+        name       : 'Сребро и естествена кожа (черна, кафява и бордо)',
+        id         : '12GRV',
+        slug       : 'srebro-i-estestvena-kozha-cherna-kafyava-i-bordo',
+        oldSlug    : '/large/large12.htm',
+        desc       : '',
+        sku        : '12GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '12grv.jpg'
+    },
+    // No19
+    {
+        price      : 189,
+        name       : 'Сребро и чемширово дърво с кожа, 2бр. кубични циркони',
+        id         : '14GRV',
+        slug       : 'srebro-i-chemshirovo-durvo-s-kozha-2br-kubichni-cirkoni',
+        oldSlug    : '/large/large14l.htm',
+        desc       : '',
+        sku        : '14GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '14grv.jpg'
+    },
+    // No19
+    {
+        price      : 179,
+        name       : 'Сребро и естествена кожа (черна, кафява и бордо)',
+        id         : '15GRV',
+        slug       : 'srebro-i-estestvena-kozha-cherna-kafyava-i-bordo',
+        oldSlug    : '/large/large15l.htm',
+        desc       : '',
+        sku        : '15GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '15grv'
+    },
+    // No20
+    {
+        price      : 198,
+        name       : 'Сребърна гривна с емайли от яспис (по поръчка се изработват с различни цветове емайли )',
+        id         : '3GRV',
+        slug       : 'sreburna-grivna-s-emajli-ot-yaspis-po-poruchka-se-izrabotvat-s-razlichni-cvetove-emajli',
+        oldSlug    : '/large/large3.htm',
+        desc       : '',
+        sku        : '3GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '3grv.jpg'
+    },
+    // No21
+    {
+        price      : 198,
+        name       : 'Ажурна сребърна гривна',
+        id         : '4GRV',
+        slug       : 'azhurna-sreburna-grivna',
+        oldSlug    : '/large/large4.htm',
+        desc       : '',
+        sku        : '4GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '4grv.jpg'
+    },
+    // No22
+    {
+        price      : 169,
+        name       : 'Сребърна гривна с малахит (по поръчка се изработват с различни камъни )',
+        id         : '5GRV',
+        slug       : 'sreburna-grivna-s-malahit-po-poruchka-se-izrabotvat-s-razlichni-kamuni',
+        oldSlug    : '/large/large5.htm',
+        desc       : '',
+        sku        : '5GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '5grv.jpg'
+    },
+    // No23
+    {
+        price      : 179,
+        name       : 'Сребро и естествена кожа (черна, кафява и бордо)',
+        id         : '6GRV',
+        slug       : 'srebro-i-estestvena-kozha-cherna-kafyava-i-bordo',
+        oldSlug    : '/large/large6.htm',
+        desc       : '',
+        sku        : '6GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '6grv.jpg'
+    },
+    // No24
+    {
+        price      : 519,
+        name       : 'Масивна сребърна гривна със златни апликации 14 к',
+        id         : '31GRV',
+        slug       : 'masivna-sreburna-grivna-sus-zlatni-aplikacii-14-k',
+        oldSlug    : '/large/31GRV.html',
+        desc       : '',
+        sku        : '31GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '31grv.jpg'
+    },
+    // No25
+    {
+        price      : 189,
+        name       : 'Сребърна гривна с чемширово дърво и кожа',
+        id         : '26GRV',
+        slug       : 'sreburna-grivna-s-chemshirovo-durvo-i-kozha',
+        oldSlug    : '/large/26grv.html',
+        desc       : '',
+        sku        : '26GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '26grv.jpg'
+    },
+    // No26
+    {
+        price      : 139,
+        name       : 'Сребърна гривна в комбинация с естествена черна кожа',
+        id         : '27GRV',
+        slug       : 'sreburna-grivna-v-kombinaciya-s-estestvena-cherna-kozha',
+        oldSlug    : '/large/27grv.html',
+        desc       : '',
+        sku        : '27GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '27grv.jpg'
+    },
+        // No27
+    {
+        price      : 128,
+        name       : 'Сребърна гривна с тюркоаз (по поръчка се изработват с различни камъни)',
+        id         : '28GRV',
+        slug       : 'sreburna-grivna-s-tyurkoaz-po-poruchka-se-izrabotvat-s-razlichni-kamuni',
+        oldSlug    : '/large/28grv.html',
+        desc       : '',
+        sku        : '28GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '28grv.jpg'
+    },
+        // No28
+    {
+        price      : 169,
+        name       : 'Сребърна гривна, изработена от въже за катерене (алпинистко)',
+        id         : '29GRV',
+        slug       : 'sreburna-grivna-izrabotena-ot-vuzhe-za-katerene-alpinistko',
+        oldSlug    : '/large/29grv.html',
+        desc       : '',
+        sku        : '27GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '29grv.jpg'
+    },
+        // No29
+    {
+        price      : 298,
+        name       : 'Ръчно изработена (кована) сребърна гривна, инкрустирана със злато 14к',
+        id         : '30GRV',
+        slug       : 'ruchno-izrabotena-kovana-sreburna-grivna-inkrustirana-sus-zlato-14k',
+        oldSlug    : '/large/30grv.html',
+        desc       : '',
+        sku        : '30GRV',
+        category   : 'sreburni grivni',
+        attributes : null,
+        image      : '30grv.jpg'
+        },
+    // No1
+    {
+        price      : 219,
+        name       : 'Ръчно изработен патиниран сребърен пръстен със златна инкрустация (14к) и брилянт (1,8mm)',
+        id         : '25PR',
+        slug       : 'ruchno-izraboten-patiniran-sreburen-prusten-sus-zlatna-inkrustaciya-14k-i-brilyant-18mm',
+        oldSlug    : '/large/rings/pr25.html',
+        desc       : '',
+        sku        : '25PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '25PR.jpg'
+    },
+    // No2
+    {
+        price      : 89,
+        name       : 'Сребърен пръстен с емайли от малахит и хематит;',
+        id         : '26PR',
+        slug       : 'sreburen-prusten-s-emajli-ot-malahit-i-hematit',
+        oldSlug    : '/large/rings/pr26.html',
+        desc       : 'Ширина на пръстена: 12 мм',
+        sku        : '26PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '26PR (1).jpg'
+    },
+    // No3
+    {
+        price      : 282,
+        name       : 'Ръчно изработен патиниран сребърен пръстен със златни апликации и инкрустиран брилянт (1,8 мм)',
+        id         : '27PR',
+        slug       : 'ruchno-izraboten-patiniran-sreburen-prusten-sus-zlatni-aplikacii-i-inkrustiran-brilyant-18-mm',
+        oldSlug    : '/large/rings/pr27.html',
+        desc       : 'Ширина на пръстена: 9 мм; Злато: 14к',
+        sku        : '27PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '27pr.jpg'
+    },
+    // No4
+    {
+        price      : 76,
+        name       : 'Ръчно изработен патиниран сребърен пръстен с класически емайли',
+        id         : '28PR',
+        slug       : 'ruchno-izraboten-patiniran-sreburen-prusten-s-klasicheski-emajli',
+        oldSlug    : '/large/rings/pr28.html',
+        desc       : '',
+        sku        : '28PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '28pr.jpg'
+    },
+    // No5
+    {
+        price      : 69,
+        name       : 'Патиниран сребърен пръстен',
+        id         : '29PR',
+        slug       : 'patiniran-sreburen-prusten',
+        oldSlug    : '/large/rings/pr29.html',
+        desc       : '',
+        sku        : '29PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '29pr.jpg'
+    },
+    // No6
+    {
+        price      : 69,
+        name       : 'Патиниран сребърен пръстен',
+        id         : '30 PR',
+        slug       : 'patiniran-sreburen-prusten',
+        oldSlug    : '/large/rings/pr30.html',
+        desc       : '',
+        sku        : '30 PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '30pr.jpg'
+    },
+    // No7
+    {
+        price      : 169,
+        name       : 'Сребърен пръстен с кубичен циркон и златна апликация -14к',
+        id         : '19PR',
+        slug       : 'sreburen-prusten-s-kubichen-cirkon-i-zlatna-aplikaciya-14k',
+        oldSlug    : '/large/rings/pr19.html',
+        desc       : 'диаметър - 38 мм',
+        sku        : '19PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '19pr.jpg'
+    },
+    // No8
+    {
+        price      : 74,
+        name       : 'Ръчно изработен сребърен пръстен с кубичен циркон',
+        id         : '20PR',
+        slug       : 'ruchno-izraboten-sreburen-prusten-s-kubichen-cirkon',
+        oldSlug    : '/large/rings/pr20.html',
+        desc       : '',
+        sku        : '20PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '20pr.jpg'
+    },
+    // No9
+    {
+        price      : 149,
+        name       : 'Масивен сребърен пръстен с кубичен циркон и емайли от хематит и яспис',
+        id         : '21PR',
+        slug       : 'masiven-sreburen-prusten-s-kubichen-cirkon-i-emajli-ot-hematit-i-yaspis',
+        oldSlug    : '/large/rings/pr21.html',
+        desc       : '',
+        sku        : '21PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '21pr.jpg'
+    },
+    // No10
+    {
+        price      : 116,
+        name       : 'Ръчно изработен сребърен пръстен Комбинация от чемширово дърво, сребро и тигрово око',
+        id         : '16PR',
+        slug       : 'ruchno-izraboten-sreburen-prusten-kombinaciya-ot-chemshirovo-durvo-srebro-i-tigrovo-oko',
+        oldSlug    : '/large/rings/pr16.htm',
+        desc       : 'Ръчно изработен сребърен пръстен Комбинация от чемширово дърво, сребро и тигрово око',
+        sku        : '16PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '16pr.jpg'
+    },
+    // No11
+    {
+        price      : 198,
+        name       : 'Ръчно изработен сребърен пръстен Комбинация от слонова кост, абанос и кехлибар (в средата)',
+        id         : '17PR',
+        slug       : 'ruchno-izraboten-sreburen-prusten-kombinaciya-ot-slonova-kost-abanos-i-kehlibar-v-sredata',
+        oldSlug    : '/large/rings/pr17.htm',
+        desc       : '',
+        sku        : '17PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '17pr.jpg'
+    },
+    // No12
+    {
+        price      : 219,
+        name       : 'Ръчно изработен сребърен пръстен със златни апликации - 14к и брилянт - 1,5 мм',
+        id         : '18PR',
+        slug       : 'ruchno-izraboten-sreburen-prusten-sus-zlatni-aplikacii-14k-i-brilyant-15-mm',
+        oldSlug    : '/large/rings/pr18.htm',
+        desc       : '',
+        sku        : '18PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '18pr.jpg'
+    },
+    // No13
+    {
+        price      : 89,
+        name       : 'Сребърен пръстен с емайли от естествени камъни - "СЕМЕЙСТВО"',
+        id         : '1PR',
+        slug       : 'sreburen-prusten-s-emajli-ot-estestveni-kamuni-semejstvo',
+        oldSlug    : '/large/rings/pr1.htm',
+        desc       : '',
+        sku        : '1PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : 'pr1.jpg'
+    },
+    // No14
+    {
+        price      : 65,
+        name       : 'Сребърен пръстен с емайли от малахит',
+        id         : '2PR',
+        slug       : 'sreburen-prusten-s-emajli-ot-malahit',
+        oldSlug    : '/large/rings/pr2.htm',
+        desc       : '',
+        sku        : '2PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : 'pr2.jpg'
+    },
+    // No15
+    {
+        price      : 79,
+        name       : 'Сребърен пръстен с емайли',
+        id         : '3PR',
+        slug       : 'sreburen-prusten-s-emajli',
+        oldSlug    : '/large/rings/pr3.htm',
+        desc       : '',
+        sku        : '3PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : 'pr3.jpg'
+    },
+    // No16
+    {
+        price      : 57,
+        name       : 'Сребърен пръстен с емайли',
+        id         : '22PR',
+        slug       : 'sreburen-prusten-s-emajli',
+        oldSlug    : '/large/rings/pr22.html',
+        desc       : '',
+        sku        : '22PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '22pr.jpg'
+    },
+    // No17
+    {
+        price      : 96,
+        name       : 'Сребърен пръстен с дърво от чемшир',
+        id         : '23PR',
+        slug       : 'sreburen-prusten-s-durvo-ot-chemshir',
+        oldSlug    : '/large/rings/pr23.html',
+        desc       : '',
+        sku        : '23PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '23pr.jpg'
+    },
+    // No18
+    {
+        price      : 139,
+        name       : 'Сребърен пръстен с въртяща се вътрешна част',
+        id         : '24PR',
+        slug       : 'sreburen-prusten-s-vurtyasha-se-vutreshna-chast',
+        oldSlug    : '/large/rings/pr24.html',
+        desc       : '',
+        sku        : '24PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '24pr.jpg'
+    },
+    // No19
+    {
+        price      : 119,
+        name       : 'Сребърен пръстен с чемшир',
+        id         : '13PR',
+        slug       : 'sreburen-prusten-s-chemshir',
+        oldSlug    : '/large/rings/pr13pr.htm',
+        desc       : '',
+        sku        : '13PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '13pr.jpg'
+    },
+    // No20
+    {
+        price      : 139,
+        name       : 'Ръчно изработени сребърни пръстени с чемширово дърво и кубични циркони',
+        id         : '14PR',
+        slug       : 'ruchno-izraboteni-sreburni-prusteni-s-chemshirovo-durvo-i-kubichni-cirkoni',
+        oldSlug    : '/large/rings/pr14.htm',
+        desc       : '',
+        sku        : '14PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : 'pr14.jpg'
+    },
+    // No21
+    {
+        price      : 129,
+        name       : 'Сребърен пръстен с емайли',
+        id         : '15PR',
+        slug       : 'sreburen-prusten-s-emajli',
+        oldSlug    : '/large/rings/pr13.htm',
+        desc       : '',
+        sku        : '15PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '15pr.jpg'
+    },
+    // No22
+    {
+        price      : 149,
+        name       : 'Сребърен пръстен',
+        id         : '4PR',
+        slug       : 'sreburen-prusten',
+        oldSlug    : '/large/rings/pr4.htm',
+        desc       : '',
+        sku        : '4PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '4pr.jpg'
+    },
+    // No23
+    {
+        price      : 89,
+        name       : 'Сребърен пръстен с емайли (по поръчка се изработват с различни цветове емайли)',
+        id         : '5PR',
+        slug       : 'sreburen-prusten-s-emajli-po-poruchka-se-izrabotvat-s-razlichni-cvetove-emajli',
+        oldSlug    : '/large/rings/pr5.htm',
+        desc       : '',
+        sku        : '5PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '5pr.jpg'
+    },
+    // No24
+    {
+        price      : 89,
+        name       : 'Сребърен пръстен',
+        id         : '6PR',
+        slug       : 'sreburen-prusten',
+        oldSlug    : '/large/rings/pr6.htm',
+        desc       : '',
+        sku        : '6PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '6pr.jpg'
+    },
+    // No25
+    {
+        price      : 96,
+        name       : 'Сребърен пръстен с емайли (по поръчка може да се изработи с различни цветове емайли)',
+        id         : '7PR',
+        slug       : 'sreburen-prusten-s-emajli-po-poruchka-mozhe-da-se-izraboti-s-razlichni-cvetove-emajli',
+        oldSlug    : '/large/rings/pr7.htm',
+        desc       : '',
+        sku        : '7PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '7pr.jpg'
+    },
+    // No26
+    {
+        price      : 98,
+        name       : 'Сребърен пръстен от две части с емайли от естествени камъни',
+        id         : '8PR',
+        slug       : 'sreburen-prusten-ot-dve-chasti-s-emajli-ot-estestveni-kamuni',
+        oldSlug    : '/large/rings/pr8.htm',
+        desc       : '',
+        sku        : '8PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '8pr.jpg'
+    },
+    // No27
+    {
+        price      : 149,
+        name       : 'Патиниран сребърен пръстен със сапфир, инкрустиран в златен обков (по поръчка може и с камък по избор)',
+        id         : '10PR',
+        slug       : 'patiniran-sreburen-prusten-sus-sapfir-inkrustiran-v-zlaten-obkov-po-poruchka-mozhe-i-s-kamuk-po-izbor',
+        oldSlug    : '/large/rings/pr10.htm',
+        desc       : '',
+        sku        : '10PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '10pr.jpg'
+    },
+    // No29
+    {
+        price      : 109,
+        name       : 'Сребърен пръстен от две частиот две части с емайли от малахит и лазурит',
+        id         : '11PR',
+        slug       : 'sreburen-prusten-ot-dve-chastiot-dve-chasti-s-emajli-ot-malahit-i-lazurit',
+        oldSlug    : '/large/rings/pr11.htm',
+        desc       : '',
+        sku        : '11PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '11pr.jpg'
+    },
+    // No30
+    {
+        price      : 89,
+        name       : 'Сребърен пръстен с емайли "СЕМЕЙСТВО"',
+        id         : '12PR',
+        slug       : 'sreburen-prusten-s-emajli-semejstvo',
+        oldSlug    : '/large/rings/pr12.htm',
+        desc       : '',
+        sku        : '12PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '12pr.jpg'
+    },
+    // No31
+    {
+        price      : 239,
+        name       : 'Сребърен пръстен (риба)',
+        id         : '31PR',
+        slug       : 'sreburen-prusten-riba',
+        oldSlug    : '/rings2/large/fishring.html',
+        desc       : 'Сребърен пръстен с черен диамант ( 0.1 k )със златен обков 14к',
+        sku        : '31PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '31pr.jpg'
+    },
+    // No32
+    {
+        price      : 169,
+        name       : 'Сребърен пръстен (халка)',
+        id         : '32PR',
+        slug       : 'sreburen-prusten-halka',
+        oldSlug    : '/large/rings/pr32.html',
+        desc       : 'Ръчно изработен сребърен пръстен, инкрустиран със злато и емайли от лазурит',
+        sku        : '32PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '32pr.jpg'
+    },
+    // No33
+    {
+        price      : 363,
+        name       : 'Сребърен пръстен с Танзанит,златен обков 18 карата',
+        id         : '33PRT',
+        slug       : 'sreburen-prusten-s-tanzanitzlaten-obkov-18-karata',
+        oldSlug    : '/large/rings/tanzanitering.html',
+        desc       : '',
+        sku        : '33PRT',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '33pr.jpg'
+    },
+    // No34
+    {
+        price      : 169,
+        name       : 'Сребърен пръстен от 2 части - син Тораз',
+        id         : '34PR',
+        slug       : 'sreburen-prusten-ot-2-chasti-sin-toraz',
+        oldSlug    : '/large/rings/2partsring.html',
+        desc       : 'Може да се изработи и с други камъни.',
+        sku        : '34PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '34pr .jpg'
+    },
+    // No35
+    {
+        price      : 194,
+        name       : 'Сребърен пръстен от 2 части с черен диамант - 0,1к',
+        id         : '35PR',
+        slug       : 'sreburen-prusten-ot-2-chasti-s-cheren-diamant-01k',
+        oldSlug    : '/large/rings/partsring2.html',
+        desc       : '',
+        sku        : '35PR',
+        category   : 'sreburni prusteni',
+        attributes : null,
+        image      : '35pr.jpg'
+    },
+    // No36
+    {
+        price      : 168,
+        name       : 'Авторски сребърни бижута',
+        id         : '36PR',
+        slug       : 'avtorski-sreburni-bizhuta',
+        oldSlug    : '/large/rings/partsring3.html',
+        desc       : 'Сребърен пръстен от 2 части със естествен Сапфир- 0,1к',
+        sku        : '36PR',
+        category   : 'sreburni prusteni',
+        attributes : 'размер 65/43 мм',
+        image      : '36pr.jpg'
+    },
+    // No1
+    {
+        price      : 169,
+        name       : 'Комплект обици и пръстен с емайли.',
+        id         : '13KM',
+        slug       : 'komplekt-obici-i-prusten-s-emajli',
+        oldSlug    : '/large/lset/set13.html',
+        desc       : 'Комплект обици и пръстен с емайли.',
+        sku        : '13KM',
+        category   : 'sreburni bijuta-komplekti',
+        attributes : null,
+        image      : '13km.jpg'
+    },
+    // No2
+    {
+        price      : 96,
+        name       : 'Сребърно колие и обици с емайли от яспис и хематит',
+        id         : '4KM',
+        slug       : 'sreburno-kolie-i-obici-s-emajli-ot-yaspis-i-hematit',
+        oldSlug    : '/large/lset/set4.htm',
+        desc       : 'Сребърно колие и обици с емайли от яспис и хематит',
+        sku        : '4KM',
+        category   : 'sreburni bijuta-komplekti',
+        attributes : null,
+        image      : '4km.jpg'
+    },
+    // No3
+    {
+        price      : 253,
+        name       : 'Сребърно колие, пръстен и обици с емайли от лазурит и малахит',
+        id         : '6 KM',
+        slug       : 'sreburno-kolie-prusten-i-obici-s-emajli-ot-lazurit-i-malahit',
+        oldSlug    : '/large/lset/set6.htm',
+        desc       : 'Сребърно колие, пръстен и обици с емайли от лазурит и малахит',
+        sku        : '6 KM',
+        category   : 'sreburni bijuta-komplekti',
+        attributes : null,
+        image      : '6km.jpg'
+    },
+    // No4
+    {
+        price      : 141,
+        name       : 'Сребърни обици и пръстен с кубични циркони',
+        id         : '12KM',
+        slug       : 'sreburni-obici-i-prusten-s-kubichni-cirkoni',
+        oldSlug    : '/large/lset/set12.html',
+        desc       : '',
+        sku        : '12KM',
+        category   : 'sreburni bijuta-komplekti',
+        attributes : null,
+        image      : '12km.jpg'
+    },
+    // No5
+    {
+        price      : 174,
+        name       : 'Сребърно колие и обици',
+        id         : '8KM',
+        slug       : 'sreburno-kolie-i-obici',
+        oldSlug    : '/large/lset/set8.htm',
+        desc       : '',
+        sku        : '8KM',
+        category   : 'sreburni bijuta-komplekti',
+        attributes : null,
+        image      : '8km.jpg'
+    },
+    // No6
+    {
+        price      : 298,
+        name       : 'Сребърно колие и пръстен в комбинация със злато - 14к и лазурит',
+        id         : '7KM',
+        slug       : 'sreburno-kolie-i-prusten-v-kombinaciya-sus-zlato-14k-i-lazurit',
+        oldSlug    : '/large/lset/set7.htm',
+        desc       : '',
+        sku        : '7KM',
+        category   : 'sreburni bijuta-komplekti',
+        attributes : null,
+        image      : '7km.jpg'
+    },
+    //No1
+    {
+        price      : 289,
+        name       : 'Ръчно изравотена сребърна врошка със златна апликация . Сребро 925 и злато 14 к',
+        id         : '9BR',
+        slug       : 'ruchno-izravotena-sreburna-vroshka-sus-zlatna-aplikaciya-srebro-925-i-zlato-14-k',
+        oldSlug    : '/large/brooch/broochfish.html',
+        desc       : 'Сребро 925 и злато 14 к размер 65/43 мм',
+        sku        : '9BR',
+        category   : 'sreburni broshki',
+        attributes : null,
+        image      : 'fish.jpg'
+    },
+    //No2
+    {
+        price      : 129,
+        name       : 'Сребърна брошка с емайли.',
+        id         : '2BR',
+        slug       : 'sreburna-broshka-s-emajli',
+        oldSlug    : '/large/brooch/brooch2.htm',
+        desc       : 'Диаметър: 5см',
+        sku        : '2BR',
+        category   : 'sreburni broshki',
+        attributes : null,
+        image      : '2br.jpg'
+    },
+    //No3
+    {
+        price      : 135,
+        name       : 'Сребърна брошка със злато 14к и емайли',
+        id         : '3BR',
+        slug       : 'sreburna-broshka-sus-zlato-14k-i-emajli',
+        oldSlug    : '/large/brooch/brooch3.htm',
+        desc       : '',
+        sku        : '3BR',
+        category   : 'sreburni broshki',
+        attributes : null,
+        image      : '3br.jpg'
+    },
+    //No4
+    {
+        price      : 968,
+        name       : 'Ръчно изработена брошка от титан и злато с брилянт - 2,2mm.',
+        id         : '5BR',
+        slug       : 'ruchno-izrabotena-broshka-ot-titan-i-zlato-s-brilyant-22mm',
+        oldSlug    : '/large/brooch/brooch5.html',
+        desc       : 'Размер: 55/45mm',
+        sku        : '5 BR',
+        category   : 'sreburni broshki',
+        attributes : null,
+        image      : '5br.jpg'
+    },
+    //No5
+    {
+        price      : 135,
+        name       : 'Сребърна брошка със златни инкрустации (14 к)',
+        id         : '6BR',
+        slug       : 'sreburna-broshka-sus-zlatni-inkrustacii-14-k',
+        oldSlug    : '/large/brooch/brooch6.html',
+        desc       : '',
+        sku        : '6BR',
+        category   : 'sreburni broshki',
+        attributes : null,
+        image      : '6br.jpg'
+    },
+    // No6
+    {
+        price      : 1039,
+        name       : 'Ръчно изработена брошка от титан и злато с брилянт (2,2mm)',
+        id         : '7BR',
+        slug       : 'ruchno-izrabotena-broshka-ot-titan-i-zlato-s-brilyant-22mm',
+        oldSlug    : '/large/brooch/brooch7.html',
+        desc       : 'диаметър: 55/45mm',
+        sku        : '7BR',
+        category   : 'sreburni broshki',
+        attributes : null,
+        image      : '7br.jpg'
+    },
+// No7
+{
+    price      : 135,
+    name       : 'Сребърна брошка с емайли.',
+    id         : '1BR',
+    slug       : 'sreburna-broshka-s-emajli',
+    oldSlug    : '/large/brooch/brooch1.htm',
+    desc       : 'Размер 5/5см',
+    sku        : '1BR',
+    category   : 'sreburni broshki',
+    attributes : null,
+    image      : '1br.jpg'
+},
+// No8
+{
+    price      : 131,
+    name       : 'Патинирана сребърна брошка с емайли',
+    id         : '8BR',
+    slug       : 'patinirana-sreburna-broshka-s-emajli',
+    oldSlug    : '/large/brooch/brooch8.html',
+    desc       : '',
+    sku        : '8BR',
+    category   : 'sreburni broshki',
+    attributes : null,
+    image      : '8br.jpg'
+},
+// No9
+{
+    price      : 118,
+    name       : 'Патинирана сребърна брошка със седеф',
+    id         : '10BR',
+    slug       : 'patinirana-sreburna-broshka-sus-sedef',
+    oldSlug    : '/large/brooch/brooch10.html',
+    desc       : 'диаметър-45мм',
+    sku        : '10BR',
+    category   : 'sreburni broshki',
+    attributes : null,
+    image      : '10br.jpg'
+},
+// No1
+{
+    price      : 159,
+    name       : 'Златен кръст (меч) с циркон.',
+    id         : '1GL',
+    slug       : 'zlaten-krust-mech-s-cirkon',
+    oldSlug    : '/Gold/cross2.html',
+    desc       : 'Размери: 31/16 мм. По поръчка може да се изработи с камък по избор: сапфир, диамант, рубин',
+    sku        : '1GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '1gl.jpg'
+},
+// No2
+{
+    price      : 599,
+    name       : 'Златен пръстен - 14к',
+    id         : '2GL',
+    slug       : 'zlaten-prusten-14k',
+    oldSlug    : '/Gold/ring%20ametist.html',
+    desc       : '- аметист ( 7 мм)   - брилянт 0.02к SI 1-F      По поръчка може да се изработи с камъни по избор',
+    sku        : '2GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '2gl.jpg'
+},
+// No3
+{
+    price      : 379,
+    name       : 'Пръстен от бяло злато(14k) брилянт VS-F (0.03k)',
+    id         : '3GL',
+    slug       : 'prusten-ot-byalo-zlato14k-brilyant-vs-f-003k',
+    oldSlug    : '/Gold/ring%20heart.html',
+    desc       : 'По поръчка може да се изработи с камък по избор',
+    sku        : '3GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '3gl.jpg'
+},
+// No4
+{
+    price      : 899,
+    name       : 'Златно колие (14к)',
+    id         : '4GL',
+    slug       : 'zlatno-kolie-14k',
+    oldSlug    : '/Gold/pendant.html',
+    desc       : '- аквамарин    - брилянти-02 k-SI1     По поръчка може да се изработи с камъни по избор',
+    sku        : '4GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '4gl.jpg'
+},
+// No5
+{
+    price      : 169,
+    name       : 'Златно кръстче (14к) с циркон',
+    id         : '5GL',
+    slug       : 'zlatno-krustche-14k-s-cirkon',
+    oldSlug    : '/Gold/cross.html',
+    desc       : 'По поръчка може да се изработи с камък по избор',
+    sku        : '5GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '5gl.jpg'
+},
+// No6
+{
+    price      : 1189,
+    name       : 'Златен пръстен от бяло и жълто злато (14к)',
+    id         : '6GL',
+    slug       : 'zlaten-prusten-ot-byalo-i-zhulto-zlato-14k',
+    oldSlug    : '/Gold/6GL.html',
+    desc       : '- с брилянт (0,25ct),SI, F/G (цената на пръстен само в един цвят, бяло или жълто злато, е с 49 лв по-ниска)',
+    sku        : '6GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '6gl.jpg'
+},
+// No7
+{
+    price      : 745,
+    name       : 'Златен пръстен от бяло и жълто злато (14к) с естествен рубин и брилянт (1,8мм)',
+    id         : '7GL',
+    slug       : 'zlaten-prusten-ot-byalo-i-zhulto-zlato-14k-s-estestven-rubin-i-brilyant-18mm',
+    oldSlug    : '/Gold/7GL.html',
+    desc       : 'По поръчка може да се изработи с камъни по избор',
+    sku        : '7GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '7gl.jpg'
+},
+// No8
+{
+    price      : 689,
+    name       : 'Златен пръстен с Аметист и брилянт-SI 1 -FG - 0.02 k',
+    id         : '16GL',
+    slug       : 'zlaten-prusten-s-ametist-i-brilyant-si-1-fg-002-k',
+    oldSlug    : '/Gold/Gold2.html',
+    desc       : '',
+    sku        : '16GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '16gl.jpg'
+},
+// No9
+{
+    price      : 729,
+    name       : 'Златен пръстен(14к) със сапфир и брилянти VS, F/G, 0,06 ct',
+    id         : '9GL',
+    slug       : 'zlaten-prusten14k-sus-sapfir-i-brilyanti-vs-fg-006-ct',
+    oldSlug    : '/Gold/9GL.html',
+    desc       : 'По поръчка може да се изработи с камък по избор',
+    sku        : '9GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '9gl.jpg'
+},
+// No10
+{
+    price      : 1280,
+    name       : 'пръстен "компас"- бяло и жълто злато 12 гр(14k)',
+    id         : '19GL',
+    slug       : 'prusten-kompas-byalo-i-zhulto-zlato-12-gr14k',
+    oldSlug    : '/Gold/19GL.html',
+    desc       : 'при въртене на малкото зъбно колело стрелката се върти- брилянти - 0,04ct и 0,02ct SI-F/G',
+    sku        : '19GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '19gl.jpg'
+},
+// No11
+{
+    price      : 1149,
+    name       : 'Златен пръстен - бяло и жълто злато (14к)',
+    id         : '11GL',
+    slug       : 'zlaten-prusten-byalo-i-zhulto-zlato-14k',
+    oldSlug    : '/Gold/11GL.html',
+    desc       : '- аквамарин 3,5ct- брилянти 10 бр. SI1, F/G (общо: 0,2ct) По поръчка може да се изработи със син топаз, аметист, цитрин.',
+    sku        : '11GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '11gl.jpg'
+},
+// No12
+{
+    price      : 899,
+    name       : 'Златен пръстен - бяло и жълто злато (14k) със сапфир и брилянти - общо 0,1ct (5 бр. по 0,02ct)',
+    id         : '12GL',
+    slug       : 'zlaten-prusten-byalo-i-zhulto-zlato-14k-sus-sapfir-i-brilyanti-obsho-01ct-5-br-po-002ct',
+    oldSlug    : '/Gold/12GL.html',
+    desc       : '- със сапфир - цена: 899 лв - с рубин -        *Цената на пръстена може да варира според качеството и вида на камъка',
+    sku        : '12GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '12gl.jpg'
+},
+// No13
+{
+    price      : 719,
+    name       : 'Златен пръстен - бяло и жълто злато (14k):',
+    id         : '13GL',
+    slug       : 'zlaten-prusten-byalo-i-zhulto-zlato-14k',
+    oldSlug    : 'Gold/13GL.html',
+    desc       : '-син топаз и брилянт 0,1ct - SI2 H - цена: 719 лв- сапфир и брилянт0,1ct- SI2 H- цена:789 лв (цената на пръстена може да варира според качеството и вида на камъка )',
+    sku        : '13GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '13gl.jpg'
+},
+// No14
+{
+    price      : 698,
+    name       : 'Златен пръстен - бяло и жълто злато (14k):',
+    id         : '14GL',
+    slug       : 'zlaten-prusten-byalo-i-zhulto-zlato-14k',
+    oldSlug    : '/Gold/14GL.html',
+    desc       : '-аметист-брилянти 2бр. - VS, F/G, 0,04ct общо  По поръчка може да се изработи със син топаз, аквамарин, турмалин.',
+    sku        : '14GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '14gl.jpg'
+},
+// No15
+{
+    price      : 525,
+    name       : 'Златна игла за вратовръзка без камък',
+    id         : '15gl',
+    slug       : 'zlatna-igla-za-vratovruzka-bez-kamuk',
+    oldSlug    : '/Gold/Pin.html',
+    desc       : 'Може да се изработи с камък по избор Крайната цена зависи от големината вида и качеството на камъка,',
+    sku        : '15gl',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '15gl.jpg'
+},
+// No16
+{
+    price      : 749,
+    name       : 'Златно колие Лилия (висулка) 3 брилянта-SI 1 -FG - 0.1k',
+    id         : '17GL',
+    slug       : 'zlatno-kolie-liliya-visulka-3-brilyanta-si-1-fg-01k',
+    oldSlug    : '/Gold/Lilia.html',
+    desc       : '25 мм-диаметър',
+    sku        : '17GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '17gl.jpg'
+},
+// No17
+{
+    price      : 769,
+    name       : 'Златен пръстен с Танзанит',
+    id         : '18GL',
+    slug       : 'zlaten-prusten-s-tanzanit',
+    oldSlug    : '/Gold/tanzanitring.html',
+    desc       : 'брилянт-SI 1 -FG - 0.02 k',
+    sku        : '18GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '18gl.jpg'
+},
+// No18
+{
+    price      : 469,
+    name       : 'Златен пръстен с Цитрин',
+    id         : '20GL',
+    slug       : 'zlaten-prusten-s-citrin',
+    oldSlug    : '/Gold/20zring.html',
+    desc       : 'брилянт-SI 1 -FG - 0.02 k',
+    sku        : '20GL',
+    category   : 'zlatni bizhuta',
+    attributes : null,
+    image      : '20gl.jpg'
+},
+// No1
+{
+    price      : 488,
+    name       : 'Ръчно изработено сребърно колие с масивен синджир (58см)',
+    id         : '5KL',
+    slug       : 'ruchno-izraboteno-sreburno-kolie-s-masiven-sindzhir-58sm',
+    oldSlug    : '/large/largekl/kl5.htm',
+    desc       : '',
+    sku        : '5KL',
+    category   : 'sreburni-kolieta',
+    attributes : null,
+    image      : '5kl.jpg'
+},
+// No9
+{
+    price      : 155,
+    name       : 'Ръчно изработено сребърно колие с чемширово дърво и кубични циркони',
+    id         : '23KL',
+    slug       : 'ruchno-izraboteno-sreburno-kolie-s-chemshirovo-durvo-i-kubichni-cirkoni',
+    oldSlug    : '/large/largekl/kl23.htm',
+    desc       : '',
+    sku        : '23KL',
+    category   : 'sreburni-kolieta',
+    attributes : null,
+    image      : '23kl.jpg'
+},
+// No9
+{
+    price      : 135,
+    name       : 'Сребърно колие с емайли от Лазурит и Яспис',
+    id         : '24KL',
+    slug       : 'sreburno-kolie-s-emajli-ot-lazurit-i-yaspis',
+    oldSlug    : '/large/largekl/kl24.htm',
+    desc       : '',
+    sku        : '24KL',
+    category   : 'sreburni-kolieta',
+    attributes : null,
+    image      : '24kl.jpeg'
+},
+// No9
+{
+    price      : 468,
+    name       : 'Сребърно колие -ръчна изработка',
+    id         : '25KL',
+    slug       : 'sreburno-kolie-ruchna-izrabotka',
+    oldSlug    : '/large/largekl/kl25.htm',
+    desc       : '',
+    sku        : '25KL',
+    category   : 'sreburni-kolieta',
+    attributes : null,
+    image      : '25kl.jpg'
+},
+// No9
+{
+    price      : 99,
+    name       : 'Сребърно колие с естествена кожа',
+    id         : '26KL',
+    slug       : 'sreburno-kolie-s-estestvena-kozha',
+    oldSlug    : '/large/largekl/kl26.htm',
+    desc       : '',
+    sku        : '26KL',
+    category   : 'sreburni-kolieta',
+    attributes : null,
+    image      : '26kl.jpg'
+},
+// No9
+{
+    price      : 189,
+    name       : 'Ръчно изработено сребърно колие',
+    id         : '27KL',
+    slug       : 'sreburno-kolie-s-estestvena-kozha',
+    oldSlug    : '/large/largekl/kl27.html',
+    desc       : '',
+    sku        : '27KL',
+    category   : 'sreburni-kolieta',
+    attributes : null,
+    image      : '27kl.jpg'
+},
+// No9
+{
+    price      : 45,
+    name       : 'Ръчно изработено сребърно колие с кожа',
+    id         : '7KL',
+    slug       : 'ruchno-izraboteno-sreburno-kolie-s-kozha',
+    oldSlug    : '/large/largekl/kl7.htm',
+    desc       : '',
+    sku        : '7KL',
+    category   : 'sreburni-kolieta',
+    attributes : null,
+    image      : '7kl.jpg'
+},
+// No9
+{
+    price      : '',
+    name       : '',
+    id         : '8KL',
+    slug       : '',
+    oldSlug    : '',
+    desc       : '',
+    sku        : '8KL',
+    category   : 'sreburni-kolieta',
+    attributes : null,
+    image      : ''
+},
+// No9
+{
+    price      : '',
+    name       : '',
+    id         : '',
+    slug       : '',
+    oldSlug    : '',
+    desc       : '',
+    sku        : '',
+    category   : 'sreburni-kolieta',
+    attributes : null,
+    image      : ''
+},
+// No9
+{
+    price      : '',
+    name       : '',
+    id         : '',
+    slug       : '',
+    oldSlug    : '',
+    desc       : '',
+    sku        : '',
+    category   : 'sreburni-kolieta',
+    attributes : null,
+    image      : ''
+},
+];
+
+const categories = [
+    {
+        name   : '',
+        id     : 'sreburni grivni',
+        slug   : '',
+        oldSlugs: ['/bangles_4.html', '/bangles_3.htm', '/bangles2.htm', '/bangles.htm', '/bangles_5.html'],
+
+    }
+]
+
+
+
+
+
